Refactor DarkModeToggle and ToggleButton components to enhance props handling and improve type definitions
