+++ src/providers/DarkThemeProvider/DarkThemeProvider.tsx
@@ -1,18 +1,17 @@
 import { type PropsWithChildren, useCallback, useMemo, useState } from 'react';
 import { LS_KEY_THEME } from '../../utils/constants';
-import { DarkThemeContext } from './DarkThemeContext';
+import { DarkThemeContext, Theme } from './DarkThemeContext';
 import { useLocalStorage } from '@src/hooks/useLocalStorage';
 
-const THEME_OPTIONS = { dark: 'dark', light: 'light' };
-
 type DarkThemeProviderProps = PropsWithChildren;
+
+const THEME_OPTIONS: { dark: Theme; light: Theme } = { dark: 'dark', light: 'light' };
 
 export default function DarkThemeProvider(props: DarkThemeProviderProps) {
   const { children } = props;
 
   const [localStorageTheme, setLocalStorageTheme] = useLocalStorage(LS_KEY_THEME);
 
-  // all useStates:
   const [isDarkMode, setIsDarkMode] = useState(() => {
     const deviceTheme = window.matchMedia('(prefers-color-scheme: dark)').matches
       ? THEME_OPTIONS.dark
@@ -27,12 +26,10 @@ export default function DarkThemeProvider(props: DarkThemeProviderProps) {
     return currentTheme === THEME_OPTIONS.dark;
   });
 
-  // all functions:
   const toggleDarkMode = useCallback(() => {
     const themeToBe = isDarkMode ? THEME_OPTIONS.light : THEME_OPTIONS.dark;
     setLocalStorageTheme(themeToBe);
 
-    // @ts-ignore
     const [htmlElement] = document.getElementsByTagName('html');
     htmlElement!.setAttribute('data-theme', themeToBe);
     document.body.setAttribute('class', themeToBe);
@@ -40,7 +37,10 @@ export default function DarkThemeProvider(props: DarkThemeProviderProps) {
     setIsDarkMode(!isDarkMode);
   }, [isDarkMode, setIsDarkMode, setLocalStorageTheme]);
 
-  const value = useMemo(() => ({ isDarkMode, toggleDarkMode }), [isDarkMode, toggleDarkMode]);
+  const value = useMemo(
+    () => ({ isDarkMode, toggleDarkMode, theme: isDarkMode ? THEME_OPTIONS.dark : THEME_OPTIONS.light }),
+    [isDarkMode, toggleDarkMode],
+  );
 
   return <DarkThemeContext.Provider value={value}>{children}</DarkThemeContext.Provider>;
 }
